WIP: move beacon types to own file
